Consolidate random state helpers
