+++ card_identifier/dataset/generator.py
@@ -12,7 +12,8 @@ import pickle
 
 from card_identifier.image import transformers, background, func_map, ImageMeta
 from card_identifier.data import get_dataset_dir, get_image_dir, get_pickle_dir
-from card_identifier.util import setup_logging, load_random_state
+from card_identifier.util import setup_logging
+from card_identifier.storage import load_random_state
 
 DEFAULT_WORKING_SIZE: Tuple[int, int] = (1024, 1024)
 DEFAULT_OUT_SIZE: Tuple[int, int] = (224, 224)
